pyproject.toml
[project]
name = "supermemory-testing-site"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "supermemory>=3.3.0",
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "anthropic>=0.39.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.0",
    "watchdog>=6.0.0",
]
